# Cache the merged config to a persistent cache