--- a/core/Core.php
+++ b/core/Core.php
@@ -61,6 +61,7 @@ gc_enable();
 
 // Include the files needed the initial manifest building, as well as any files
 // that are needed for the boostrap process on every request.
+require_once 'core/ConfigCacheInterface.php';
 require_once 'cache/Cache.php';
 require_once 'core/Object.php';
 require_once 'core/ClassInfo.php';