This caches merged config to a persistent cache so that it doesn't have to merge on each request.

Initial tests show page load times on a base installation drop from ~~2 seconds to 0.35 seconds~~ (edit: not properly performance tested).

I'll provide more data shortly.

**Key**
localhost:8181 - **master**
vagrant:8181 - **pre-merge-config**