@kinglozzer 2015-11-06
Interesting approach of disabling the persistent cache for certain items when `Config::inst()->update()` is called. First question (always seems to be the case with these changes!) - how is memory use impacted?

If a request to a home page adds 50 items to the persistent cache, then a subsequent request to a different page adds another 50, you end up with 100 items in the cache: requests after that will then load all 100 items into memory. Eventually we’ll end up with nearly every config setting being effectively pre-loaded into memory, right?