@micmania1 2015-11-06
I'm continuing work on this today. I expect some memory increase but not huge. The private statics change I made saved ~1mb so I'm cashing that in :)