@micmania1 2015-11-07
Okay, so I've had a chance to performance test it today. There's a decent speed increase, but not 2 to 0.3 seconds. It drops from ~650ms to ~580ms (10%).

![xhprof-1](https://cloud.githubusercontent.com/assets/881537/11012350/70281a8e-8557-11e5-9b0e-de36956fb58e.png)

The calls to SS_ConfigStaticManifest and SS_ConfigManifest drop from ~1100 to a total of 8 which is where the savings come from.

Another point to note is that memory actually drops with this PR. I'm assuming the reason for this is that its loading the config from a single file, rather than loading from 100s of files. I'm assuming there's overhead in doing this.

These tests were on a base install, MariaDB 10.0, PHP 5.6 and Apache 2.4.

I'll test on a real project and see whether there's any gains or not. However, they'll be on 3.2 and a cherry-picked commit as I don't know any projects running on 4.0.