@micmania1 2015-11-07
For this test I used the silverstripe demo project (one ring rentals site) with demo content. My version is running 3.2.0 with this commit cherry-picked.

![xhprof-2](https://cloud.githubusercontent.com/assets/881537/11013637/4e902cae-857a-11e5-8bf1-909edef04d60.png)