@micmania1 2015-11-07
https://github.com/silverstripe/silverstripe-framework/issues/4752 - I've raised this as a bug.

Its to do with the registerChangeCallback method which is removed in this PR. I don't think removing it is sensible as this means changing dev mode for example, won't clear the cache. Its important that it does clear the cache because config is different depending on except/only blocks in yaml config.