@micmania1 2015-12-04
I've refactored this a bit from my first PR and included @sminnee fix from #4762

I need to do some more performance testing to make sure I haven't introduced anything strange but then I'll remove the DNM tag and squash when/if everyone is happy.