@micmania1 2016-01-17
I've closed this for now but will re-open. I'm getting really up and down results on performance tests so will figure out why.